+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Use the cached graph analysis of `ewokscore` (`TaskGraph.analysis`), which makes building
+  a workflow from a large graph faster.
+
 ## [3.1.0rc1] - 2026-08-04
 
 ### Added